+++ src/lib/DatePickerInput.tsx
@@ -18,7 +18,7 @@ interface DatePickerInputProps<T extends FieldValues> extends Omit<CommonInputPr
   /**
    * The props for the date picker component: https://reactdatepicker.com/
    */
-  datePickerProps?: Omit<ReactDatePickerProps, "onChange" | "selected" | "id" | "className" | "onBlur">;
+  datePickerProps?: Omit<ReactDatePickerProps, "onChange" | "selected" | "id" | "className" | "onBlur" | "autoComplete">;
 
   /**
    * The onChange handler for the date picker component.
@@ -38,6 +38,11 @@ interface DatePickerInputProps<T extends FieldValues> extends Omit<CommonInputPr
    * The ref of the date picker component.
    */
   datePickerRef?: MutableRefObject<DatePicker<never, undefined> | null>;
+
+  /**
+   * The autoComplete property for the date picker component.
+   */
+  autoComplete?: string;
 }
 
 const DEFAULT_DATE_FORMAT = "dd.MM.yyyy";
@@ -54,6 +59,7 @@ const DatePickerInput = <T extends FieldValues>(props: DatePickerInputProps<T>) 
     addonRight,
     ianaTimeZone,
     className = "",
+    autoComplete = "off",
     inputGroupStyle,
     datePickerRef,
     name: initialName,
@@ -144,6 +150,7 @@ const DatePickerInput = <T extends FieldValues>(props: DatePickerInputProps<T>) 
             dateFormat={effectiveDateFormat}
             calendarStartDay={calendarStartDay}
             wrapperClassName={error ? "is-invalid" : ""}
+            autoComplete={autoComplete}
             selected={selectedDate}
             ref={(elem) => {
               // https://github.com/react-hook-form/react-hook-form/discussions/5413